convert param _index from int to unsigned int

+++ gazebo/physics/bullet/BulletJoint.cc
@@ -367,7 +367,8 @@ void BulletJoint::SetDamping(int _index, double _damping)
 {
   if (static_cast<unsigned int>(_index) < this->GetAngleCount())
   {
-    this->SetStiffnessDamping(_index, this->stiffnessCoefficient[_index],
+    this->SetStiffnessDamping(static_cast<unsigned int>(_index),
+      this->stiffnessCoefficient[_index],
       _damping);
   }
   else
@@ -380,10 +381,10 @@ void BulletJoint::SetDamping(int _index, double _damping)
 }
 
 //////////////////////////////////////////////////
-void BulletJoint::SetStiffnessDamping(int _index,
+void BulletJoint::SetStiffnessDamping(unsigned int _index,
   double _stiffness, double _damping, double _reference)
 {
-  if (_index < static_cast<int>(this->GetAngleCount()))
+  if (_index < this->GetAngleCount())
   {
     this->stiffnessCoefficient[_index] = _stiffness;
     this->dampingCoefficient[_index] = _damping;
